Cleanup AddWallet components
- Removed dead code
- Moved intersperse utility to ui/utils/lists
- Refactored styles
